+++ admin/index.php
@@ -56,6 +56,7 @@ if ($administrator->isAuthorized() && $core->detectAdminMode() == 'login') {
 if (!$administrator->isLogged() && $core->detectAdminMode() != 'lostpwd')
     include_once('login.php');
 elseif ($core->detectAdminMode() == 'plugin') {
+    $core->callHook('adminBeforeRunPlugin');
     if ($runPlugin->getAdminTemplate()) {
         if (util::getFileExtension($runPlugin->getAdminTemplate()) === 'tpl') {
             $layout = new Template(ADMIN_PATH . 'layout.tpl');